Removes root from Dialog
