+++ packages/react-components/react-dialog/src/components/Dialog/Dialog.tsx
@@ -4,10 +4,9 @@ import { renderDialog_unstable } from './renderDialog';
 import { useDialogStyles_unstable } from './useDialogStyles';
 import type { DialogProps } from './Dialog.types';
 import { useDialogContextValues_unstable } from './useDialogContextValues';
-import { ForwardRefComponent } from '@fluentui/react-utilities';
 
-export const Dialog: ForwardRefComponent<DialogProps> = React.forwardRef((props, ref) => {
-  const state = useDialog_unstable(props, ref);
+export const Dialog: React.FC<DialogProps> = React.memo(props => {
+  const state = useDialog_unstable(props);
   const contextValues = useDialogContextValues_unstable(state);
 
   useDialogStyles_unstable(state);
