+++ packages/react-components/react-dialog/src/components/DialogContent/useDialogContent.ts
@@ -1,8 +1,9 @@
 import * as React from 'react';
-import { getNativeElementProps, useMergedRefs } from '@fluentui/react-utilities';
+import { getNativeElementProps, useEventCallback, useMergedRefs } from '@fluentui/react-utilities';
 import type { DialogContentProps, DialogContentState } from './DialogContent.types';
 import { useDialogContext_unstable } from '../../contexts/dialogContext';
 import { useModalAttributes } from '@fluentui/react-tabster';
+import { isEscapeKeyDismiss } from '../../utils/isEscapeKeyDown';
 
 /**
  * Create the state required to render DialogContent.
@@ -20,10 +21,19 @@ export const useDialogContent_unstable = (
   const { as = 'div' } = props;
 
   const contentRef = useDialogContext_unstable(ctx => ctx.contentRef);
-  const type = useDialogContext_unstable(ctx => ctx.modalType);
+  const requestOpenChange = useDialogContext_unstable(ctx => ctx.requestOpenChange);
+  const modalType = useDialogContext_unstable(ctx => ctx.modalType);
 
   const { modalAttributes } = useModalAttributes({
-    trapFocus: type !== 'non-modal',
+    trapFocus: modalType !== 'non-modal',
+  });
+
+  const handleRootKeyDown = useEventCallback((event: React.KeyboardEvent<HTMLDivElement>) => {
+    props.onKeyDown?.(event);
+    if (isEscapeKeyDismiss(event, modalType)) {
+      requestOpenChange({ event, open: false, type: 'escapeKeyDown' });
+      event.preventDefault();
+    }
   });
 
   return {
@@ -34,6 +44,7 @@ export const useDialogContent_unstable = (
       ref: useMergedRefs(ref, contentRef),
       ...props,
       ...modalAttributes,
+      onKeyDown: handleRootKeyDown,
     }),
   };
 };
